Clean dossier with only the changes of the guestbook and one php page dedicated only to the guestbook, changed the polices, images and css associated to the guestbook.
