--- css/template.css
@@ -1,8 +1,0 @@
-.overdue {
-  color: red;
-}
-
-.done {
-  color: grey;
-  text-decoration: line-through;
-}
