--- footer.php
@@ -1,26 +1,0 @@
-<footer class="page-footer font-small orange">
-    <div class="container">
-        <ul class="row text-center d-flex justify-content-center pt-5 mb-3">
-            <li class="col-xs-12 col-sm-2 text-uppercase font-weight-bold">
-                    <a href="faq.php">F.A.Q.</a>
-            </li>
-            <li class="col-xs-12 col-sm-3 text-uppercase font-weight-bold">
-                    <a href="mention.php">Legal Mentions</a>
-            </li>
-            <li class="col-xs-12 col-sm-3 text-uppercase font-weight-bold">
-                    <a href="index.php#contact">Contact</a>
-            </li>
-            <li class="col-xs-12 col-sm-2 text-uppercase font-weight-bold">
-                <img class="high" src="img/facebook.png" alt="facebook">
-            </li>
-            <li class="col-xs-12 col-sm-2 text-uppercase font-weight-bold">
-                <img class="high" src="img/twitter.png" alt="Twitter">
-            </li>
-        </ul>
-        <div class="row">
-            <div class="text-center py-3">
-                <a href="https://wildcodeschool.fr/" target="_blank">Made with love by Wild Code School</a>
-            </div>
-        </div>
-    </div>
-</footer>
