--- header.php
@@ -1,27 +1,0 @@
-<header>
-    <nav class="navbar navbar-expand-lg orange">
-        <a class="navbar-brand" href="index.php">Wild Code School</a>
-        <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
-            <img class="navbar-toggler-icon" src="img/button.jpg">
-        </button>
-        <div class="collapse navbar-collapse" id="navbarNav">
-            <ul class="navbar-nav">
-                <li class="nav-item active">
-                    <a class="nav-link" href="index.php">Home <span class="sr-only">(current)</span></a>
-                </li>
-                <li class="nav-item">
-                    <a class="nav-link" href="students.php">Students</a>
-                </li>
-                <li class="nav-item">
-                    <a class="nav-link" href="trainers.php">Trainers</a>
-                </li>
-                <li class="nav-item">
-                    <a class="nav-link" href="joys.php">The Joys of the Wild</a>
-                </li>
-                <li class="nav-item">
-                    <a class="nav-link" href="euratech.php">Euratechnologies</a>
-                </li>
-            </ul>
-        </div>
-    </nav>
-</header>
