--- index.php
@@ -1,22 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <link rel="stylesheet" href="css/template.css">
-    <title>Document</title>
-</head>
-
-<body>
-    <h1 id="test">Test</h1>
-    <ul id="todoList">
-        <!-- Tasks come here ! -->
-    </ul>
-    <button id="addTaskButton">Add a Task</button>
-
-    <script src="guestbook.js"></script>
-</body>
-
-</html>
