--- trainers.php
@@ -1,70 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-    <meta name="description" content="Fun website of the Wild Code School">
-  	<meta name="keywords" content="HTML,CSS,PHP,JavaScript, Wild Code School">
-  	<meta name="authors" content="Nicolas Duhamel, Emilie Boeglen, Julie Dulmas, Steven Antal, Arlensiú Celis">
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css" integrity="sha384-MCw98/SFnGE8fJT3GXwEOngsV7Zt27NXFoaoApmYm81iuXoPkFOJwJ8ERdknLPMO" crossorigin="anonymous">
-    <link rel="stylesheet" href="css/template.css">
-    <title>WCS - Trainers</title>
-</head>
-
-<body>
-<?php require('header.php'); ?>
-<section id="bannerfluid">
-	<div class="container-fluid">
-    	<img class="d-block w-100 pb-3" src="https://via.placeholder.com/350x75" alt="Trainers">
-	</div>
-</section>
-
-<section id="trainers">
-	<div class="container">
-	    <div class="row justify-content-center">
-	        <div class="col-8">
-	            <div class="media">
-	                <img class="backform align-self-center" src="https://assets.pokemon.com/assets/cms2/img/pokedex/full/055.png" alt="formateur1">
-	                <img class="frontform align-self-center" src="https://assets.pokemon.com/assets/cms2/img/pokedex/full/054.png" alt="egoformateur1">
-	                <div class="media-body align-self-center py-3 ml-4 px-8 text-justify">
-	                    <h5 class="mt-0 mb-3">Trainer 1 - Title</h5>
-	                    <p>Cras sit amet nibh libero, in gravida nulla. Nulla vel metus scelerisque ante sollicitudin. Cras purus odio, vestibulum in vulputate at, tempus viverra turpis. Fusce condimentum nunc ac nisi vulputate fringilla. Donec lacinia congue felis in faucibus.</p>
-	                </div>
-	            </div>
-	        </div>
-
-	        <div class="col-8">
-	            <div class="media">
-	                <div class="media-body align-self-center py-3 mr-4 px-8 text-justify">
-	                    <h5 class="mt-0 mb-3">Trainer 2 - Title</h5>
-	                    <p>Cras sit amet nibh libero, in gravida nulla. Nulla vel metus scelerisque ante sollicitudin. Cras purus odio, vestibulum in vulputate at, tempus viverra turpis. Fusce condimentum nunc ac nisi vulputate fringilla. Donec lacinia congue felis in faucibus.</p>
-	                </div>
-	                <img class="backform align-self-center" src="https://assets.pokemon.com/assets/cms2/img/pokedex/full/055.png" alt="formateur1">
-	                <img class="frontform align-self-center" src="https://assets.pokemon.com/assets/cms2/img/pokedex/full/054.png" alt="egoformateur1">
-	            </div>
-	        </div>
-
-	        <div class="col-8">
-	            <div class="media">
-	                <img class="backform align-self-center" src="https://assets.pokemon.com/assets/cms2/img/pokedex/full/055.png" alt="formateur1">
-	                <img class="frontform align-self-center" src="https://assets.pokemon.com/assets/cms2/img/pokedex/full/054.png" alt="egoformateur1">
-	                <div class="media-body align-self-center py-3 ml-4 px-8 text-justify">
-	                    <h5 class="mt-0 mb-3">Trainer 3 - Title</h5>
-	                    <p>Cras sit amet nibh libero, in gravida nulla. Nulla vel metus scelerisque ante sollicitudin. Cras purus odio, vestibulum in vulputate at, tempus viverra turpis. Fusce condimentum nunc ac nisi vulputate fringilla. Donec lacinia congue felis in faucibus.</p>
-	                </div>
-	            </div>
-	        </div>
-	    </div>
-	</div>
-</section>
-<?php require('footer.php'); ?>
-
-
-<!-- jQuery first, then Tether, then Bootstrap JS. -->
-<script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous"></script>
-<script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.3/umd/popper.min.js" integrity="sha384-ZMP7rVo3mIykV+2+9J3UJ46jBk0WLaUAdn689aCwoqbBJiSnjAK/l8WvCWPIPm49" crossorigin="anonymous"></script>
-<!--Old script (popper) that did not appear in the new documentation that I just put here between-->
-<script src="https://cdnjs.cloudflare.com/ajax/libs/tether/1.4.0/js/tether.min.js" integrity="sha384-DztdAPBWPRXSA/3eYEEUWrWCy7G5KFbe8fFjk5JAIxUYHKkDx6Qin1DkWx51bBrb" crossorigin="anonymous"></script>
-<script src="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/js/bootstrap.min.js" integrity="sha384-ChfqqxuZUCnJSK3+MXmPNIyE6ZbWh2IMqE241rYiqJxyMiZ6OW/JmZQ5stwEULTy" crossorigin="anonymous"></script>
-</body>
-</html>
